Extract common ControllerExceptionHandler to use in all our apis, in order to handle system error, NotFound and spring security AccessDeniedException
