+++ applications/registration-api/src/main/java/no/dcat/controller/CatalogController.java
@@ -82,7 +82,6 @@ public class CatalogController {
                                                             PagedResourcesAssembler assembler) {
 
         Authentication auth = springSecurityContextBean.getAuthentication();
-
         Set<String> validCatalogs = new HashSet<>();
 
         for (GrantedAuthority authority : auth.getAuthorities()) {
