+++ applications/registration-api/src/main/java/no/dcat/controller/ControllerExceptionHandler.java
@@ -1,43 +1,8 @@
 package no.dcat.controller;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.AccessDeniedException;
+import no.dcat.webutils.exceptions.GlobalControllerExceptionHandler;
 import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.context.request.WebRequest;
-import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
 
-import java.util.HashMap;
-import java.util.Map;
-
-/**
- * Created by bjg on 04.07.2017.
- */
 @ControllerAdvice
-public class ControllerExceptionHandler {
-    private static Logger logger = LoggerFactory.getLogger(ControllerExceptionHandler.class);
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
-    @ResponseBody
-    protected Object handleException(Exception ex) {
-        logger.warn("Unable to handle service call ",ex);
-        Map<String,String> result = new HashMap<>();
-        result.put("error", ex.getLocalizedMessage());
-        return result;
-    }
-
-    @ResponseStatus(HttpStatus.FORBIDDEN)
-    @ExceptionHandler({AccessDeniedException.class})
-    @ResponseBody
-    public Object accessDenied(Exception ex) {
-        Map<String,String> result = new HashMap<>();
-        result.put("error", ex.getLocalizedMessage());
-        return result;
-    }
+public class ControllerExceptionHandler extends GlobalControllerExceptionHandler {
 }
