+++ libraries/webutils-spring1/src/main/java/no/dcat/webutils/exceptions/NotFoundException.java
@@ -1,0 +1,12 @@
+package no.dcat.webutils.exceptions;
+
+public class NotFoundException extends Exception {
+
+    public NotFoundException() {
+    }
+
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
+
